Fonction 2004 (Se désister) Fonction 2006 (Annuler)

+++ src/Controller/SortieController.php
@@ -34,6 +34,10 @@ class SortieController extends AbstractController
 
         $data = new RechercheData();
 
+
+        if ($data->campus == null){
+            $data->campus = $campusRepository->find($this->getUser()->getCampus());
+        }
         $form = $this->createForm(RechercheSortieType::class, $data);
 
         $form->handleRequest($request);
@@ -94,6 +98,7 @@ class SortieController extends AbstractController
      */
     public function show(Sortie $sortie, $action = 0): Response
     {
+
         if($action) {
             switch ($action){
                 case 1:
@@ -104,6 +109,10 @@ class SortieController extends AbstractController
                     break;
                 case 3:
                     $this->actionAnnuler($sortie);
+
+                    break;
+                default:
+
                     break;
             }
 
@@ -140,15 +149,18 @@ class SortieController extends AbstractController
      */
     public function delete(Request $request, Sortie $sortie): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$sortie->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($sortie);
-            $entityManager->flush();
-        }
+       if ($sortie->getOrganisateur() == $this->getUser()) {
+           if ($this->isCsrfTokenValid('delete'.$sortie->getId(), $request->request->get('_token'))) {
+               $entityManager = $this->getDoctrine()->getManager();
+               $entityManager->remove($sortie);
+               $entityManager->flush();
+           }
+       }
 
         return $this->redirectToRoute('sortie_index');
     }
 
+    // Récupère le nombre d'inscrits par sortie et vérifie que notre User est inscrit
     public function getInscriptions($sorties) {
         foreach ($sorties as $sortie) {
             $participants = $sortie->getParticipants();
@@ -164,179 +176,22 @@ class SortieController extends AbstractController
         return $sorties;
     }
 
-    public function findByInscrit($data, ParticipantRepository $participantRepository) {
-        $SortiesOfUser = null;
 
-        if (!empty($data->user)) {
-            $participant = $participantRepository->find($data->user);
-            dd($participant);
-        }
-    }
+    // Contrôle de la date limite de clôture des inscriptions + si déjà inscrit
+    public function actionSeDesister($sortie) {
 
-    private function actionSInscrire(Sortie $sortie){
-
-        if($sortie->getEtat()->getId() == 2){
-            $user = $this->getUser();
-
-            //en premier lieu on vérifie l'etat de la sortie qui doit etre ouverte pour pouvoir s'inscrire
-            if($sortie->getEtat() == 2){
-
-                //on vérifie que l'user n'est pas déjà inscrit
-                if(!$sortie->getParticipants()->contains($user))
-                {
-                    if($sortie->getOrganisateur()->getId() == $user->getId())
-                    {
-                        $message = "Vous êtes l'organisateur de cette sortie : " . $sortie->getNom();
-                        $this->addFlash("warning", $message);
-                    }
-                    else{
-
-                        //on vérifie le nombre d'inscrit
-                        //normalement pas de lien vers l'inscription
-                        if($sortie->getParticipants()->count() < $sortie->getNbInscriptionsMax()){
-
-                            $sortie->addParticipant($user);
-
-                            //si participants = nombre max on cloture la sortie
-                            if($sortie->getParticipants()->count() == $sortie->getNbInscriptionsMax()){
-
-                                $sortie->setEtat(3);
-                            }
-
-                            $this->getDoctrine()->getManager()->flush();
-
-                            $message = "Vous êtes maintenant inscrit à la sortie : " . $sortie->getNom();
-                            $this->addFlash("success", $message);
-
-                        }else{
-
-                            $message = "La sortie : '" . $sortie->getNom() . "' a déjà atteint son nombre de participants maximum.";
-                            $this->addFlash("danger", $message);
-                        }
-                    }
-                }
-                else{
-
-                    $message = "Vous êtes déjà inscrit à cette sortie : " . $sortie->getNom();
-                    $this->addFlash("danger", $message);
-                }
-            }
-        }
-        return $this->redirectToRoute('sortie_index');
-    }
-
-    private function actionAnnuler(Sortie $sortie)
-    {
-    }
-
-    private function actionSeDesister(Sortie $sortie)
-    {
-    }
-
-    //methode qui modifie l'etat d'une sortie pour l'archiver selon certaines conditions
-    private function archiverLesSorties(SortieRepository $sortieRepository, EtatRepository $etatRepository){
-
-        $sorties = $sortieRepository->findAll();
-
-        $today = date_create('now');
-
-        foreach($sorties as $sortie) {
-
-            //on archive les sorties terminées ou annulées après un mois
-            if ($sortie->getEtat()->getId() == 5 || $sortie->getEtat()->getId() == 6) {
-
-                $dateSortie = date_create($sortie->getDateHeureDebut()->format('Y-m-d'));
-                $datePlus1Month = date_add($dateSortie, date_interval_create_from_date_string('1 month'));
-
-                if ($datePlus1Month < $today) {
-                    $sortie->setEtat($etatRepository->find(7));
+        $participants = $sortie->getParticipants();
+        if ($sortie->getDateLimiteInscription() >= date_create('now')->format('Y-m-d H:i:s')) {
+            foreach ($participants as $part) {
+                if ($part == $this->getUser() ) {
+                    $sortie->removeParticipant($this->getUser());
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->persist($sortie);
+                    $entityManager->flush();
                 }
             }
         }
 
-        $this->getDoctrine()->getManager()->flush();
-
-        return $this->redirectToRoute('sortie_index');
-    }
-
-    //methode qui modifie l'etat d'une sortie pour la cloturer selon certaines conditions
-    private function cloturerLesSorties(SortieRepository $sortieRepository, EtatRepository $etatRepository){
-
-        $sorties = $sortieRepository->findAll();
-
-        $today = date_create('now');
-
-        foreach($sorties as $sortie)
-        {
-            //on cloture les sorties ouvertes
-            if($sortie->getEtat()->getId() == 2){
-
-                //on vérifie pour chaque sortie le nombre d'inscrits
-                //normalement vérification faite lors de l'inscription
-                if($sortie->getParticipants()->count() == $sortie->getNbInscriptionsMax())
-                {
-                    $sortie->setEtat($etatRepository->find(3));
-                }
-
-                //si le nombre d'inscrits est inférieur au nombre max on vérifie la date limite
-                if($sortie->getDateLimiteInscription() < $today )
-                {
-                    $sortie->setEtat($etatRepository->find(3));
-                }
-            }
-        }
-
-        $this->getDoctrine()->getManager()->flush();
-
-        return $this->redirectToRoute('sortie_index');
-    }
-
-    private function MettreAJourActivite(SortieRepository $sortieRepository, EtatRepository $etatRepository){
-
-        //TODO à finir !! vérifier les comparaisons de date
-
-        $sorties = $sortieRepository->findAll();
-
-        $maintenant = new DateTime('now');
-        $today = date_create((new DateTime('now'))->format('Y-m-d'));
-
-        foreach($sorties as $sortie) {
-
-            //pour qu'on modifie l'activité à en cours ou terminée il faut vérifier l'etat
-            //la sortie doit être à ouverte ou cloturée
-            if ($sortie->getEtat()->getId() == 2 || $sortie->getEtat()->getId() == 3) {
-
-                $jourSortie  = date_create($sortie->getDateHeureDebut()->format('Y-m-d'));
-                $heureSortie = date_create($sortie->getDateHeureDebut()->format('Y-m-d HH:mm'));
-
-                //on compare la date du jour avec la date de sortie
-                if ( $jourSortie == $today && $heureSortie < $maintenant){
-
-                    //sortie en cours
-                    $sortie->setEtat($etatRepository->find(4));
-                }
-            }
-
-            //l'activité ne doit ni être annulée ni archivée ni non publiée
-            //pour passer à activité terminée si la date de l'activité est passée
-            if($sortie->getEtat() != 6 && $sortie->getEtat() != 7 && $sortie->getEtat() != 1) {
-
-                $jourSortie  = date_create($sortie->getDateHeureDebut()->format('Y-m-d'));
-                $debutSortie = date_create($sortie->getDateHeureDebut()->format('Y-m-d HH:mm'));
-
-                $intervalString = "PT" . $sortie->getDuree(). "M";
-                $duree = new DateInterval($intervalString);
-
-                $heureFinSortie = date_add($debutSortie,$duree);
-
-                if ($jourSortie < $today || ($jourSortie == $today && $heureFinSortie < $maintenant)){
-
-                    //sortie en cours
-                    $sortie->setEtat($etatRepository->find(5));
-                }
-
-            }
-        }
     }
 
 }
